convert to defensivebot2
