+++ .gitignore
@@ -21,7 +21,7 @@ build/*
 *.iws
 /.idea/
 
-/src/defensivebot2
+
 
 # User-specific stuff:
 battlecode-scaffold.iws
