+++ src/defensivebot2/bots/Miner.java
@@ -1,11 +1,11 @@
-package defensivebot.bots;
+package defensivebot2.bots;
 
 
 import battlecode.common.*;
-import defensivebot.utils.Constants;
+import defensivebot2.utils.Constants;
 
-import static defensivebot.bots.Archon.rng;
-import static defensivebot.utils.Constants.directions;
+import static defensivebot2.bots.Archon.rng;
+import static defensivebot2.utils.Constants.directions;
 
 public class Miner extends Robot{
