+++ src/defensivebot2/bots/Sage.java
@@ -1,10 +1,10 @@
-package defensivebot.bots;
+package defensivebot2.bots;
 
 import battlecode.common.GameActionException;
 import battlecode.common.RobotController;
 
-public class Lab extends Robot{
-    public Lab(RobotController rc) throws GameActionException  {
+public class Sage extends Robot{
+    public Sage(RobotController rc) throws GameActionException  {
         super(rc);
     }
