+++ src/defensivebot2/bots/WatchTower.java
@@ -1,16 +1,16 @@
-package defensivebot.bots;
+package defensivebot2.bots;
 
 import battlecode.common.GameActionException;
 import battlecode.common.RobotController;
 
-public class Sage extends Robot{
-    public Sage(RobotController rc) throws GameActionException  {
+public class WatchTower extends Robot{
+    public WatchTower(RobotController rc) throws GameActionException  {
         super(rc);
     }
 
     @Override
     public void sense() throws GameActionException {
-
+        
     }
 
     @Override
