+++ src/defensivebot2/datasturctures/LinkedList.java
@@ -1,4 +1,4 @@
-package defensivebot.datasturctures;
+package defensivebot2.datasturctures;
 
 public class LinkedList<T> {
     public int size = 0;
