+++ src/defensivebot2/datasturctures/Node.java
@@ -1,4 +1,4 @@
-package defensivebot.datasturctures;
+package defensivebot2.datasturctures;
 
 public class Node<T> {
     public Node<T> next;
