+++ src/defensivebot2/enums/CommInfoBlockType.java
@@ -1,4 +1,4 @@
-package defensivebot.enums;
+package defensivebot2.enums;
 
 public enum CommInfoBlockType {
