+++ src/defensivebot2/utils/LogUtils.java
@@ -1,8 +1,8 @@
-package defensivebot.utils;
+package defensivebot2.utils;
 
 import battlecode.common.RobotController;
 
-import static defensivebot.bots.Robot.turnCount;
+import static defensivebot2.bots.Robot.turnCount;
 
 public class LogUtils {
